Add Typing Override from ZHAO-II-MB2.x
This is a simple port of the necessary code to make a typing override
work. A better fix would be to rebuild the Opencollar AO on top of the
MB2.x modified ZHAO, which includes memory optimizations, but I am not
sure how the OC devs feel about doing that. I did my best to mimic the
style of the original script, though IMO parts of it are horrible to try
and read.